Move LoRA related things to a LoraUtil (have to think about where to best place this later...)
